+++ libraries/DLTest/DLTest.Mock.delay.h
@@ -1,6 +1,6 @@
 #ifndef _DL_TEST_MOCK_DELAY_H_
 #define _DL_TEST_MOCK_DELAY_H_
 
-void delay(uint8_t);
+void delay(long);
 
 #endif
